a lot of fixes
git-svn-id: file:///usr/local/svn/cpplib/trunk/cftal@1724 18305beb-b1e2-0310-bcd8-b14b5419b1d7

+++ include/cftal/vec_t_1.h
@@ -204,9 +204,9 @@ namespace cftal {
         struct mask_to_bool<float> {
             static
             bool
-            v(double t) {
+            v(float t) {
                 int32_t ti=as<int32_t>(t);
-                return mask_to_bool<int64_t>::v(ti);
+                return mask_to_bool<int32_t>::v(ti);
             }
         };
